refactor: update OrderApi imports to use the services directory
