+++ src/app/features/checkout/components/checkout-review-step/checkout-review-step.spec.ts
@@ -6,7 +6,7 @@ import { CheckoutReviewStep } from './checkout-review-step';
 import { CheckoutWizard } from '../../services/checkout-wizard';
 import { checkoutRoutes } from '../../checkout.routes';
 import { CartStore } from '../../../cart/cart.store';
-import { OrderApi } from '../../../orders/order-api';
+import { OrderApi } from '../../../orders/services/order-api';
 
 const testRoutes: Routes = [{ path: 'checkout', children: checkoutRoutes }];
