+++ src/app/features/checkout/components/checkout-schedule-step/checkout-schedule-step.spec.ts
@@ -6,7 +6,7 @@ import { CheckoutScheduleStep } from './checkout-schedule-step';
 import { CheckoutWizard } from '../../services/checkout-wizard';
 import { checkoutRoutes } from '../../checkout.routes';
 import { CartStore, CartItem, CartData } from '../../../cart/cart.store';
-import { OrderApi } from '../../../orders/order-api';
+import { OrderApi } from '../../../orders/services/order-api';
 
 const cartStoreStub = {
   totalPrice: signal(0),
