+++ src/app/features/checkout/guards/checkout-step.guard.spec.ts
@@ -6,7 +6,7 @@ import { checkoutStepGuard } from './checkout-step.guard';
 import { CheckoutWizard } from '../services/checkout-wizard';
 import { checkoutRoutes } from '../checkout.routes';
 import { CartStore, CartItem, CartData } from '../../cart/cart.store';
-import { OrderApi } from '../../orders/order-api';
+import { OrderApi } from '../../orders/services/order-api';
 
 const cartStoreStub = {
   totalPrice: signal(0),
