+++ src/app/features/checkout/services/checkout-wizard.spec.ts
@@ -5,7 +5,7 @@ import { describe, it, expect, beforeEach, vi } from 'vitest';
 import { CheckoutWizard } from './checkout-wizard';
 import { checkoutRoutes } from '../checkout.routes';
 import { CartStore, CartItem, CartData } from '../../cart/cart.store';
-import { OrderApi } from '../../orders/order-api';
+import { OrderApi } from '../../orders/services/order-api';
 
 const cartStoreStub = {
   totalPrice: signal(0),
